requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.10.3
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
django-bootstrap-pagination==1.7.1
django-filter==2.4.0
djangorestframework==3.10.3
schedule==0.6.0

